solved 17.2
almost no code had to change!!!  yay!!!  this was a very easy part2.  just attach a singleton dimension to the front of the state array in dimension 3, and a full dimension to the front in dimension 4.  it was pleasant that we didn't have to change the rules for when a cube activates.

the hardest part of this problem remains understanding the given example for 3d.